--- client/src/components/Core/Breadcrumb/Breadcrumbs.module.scss
@@ -1,4 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: row;
-}
