+++ client/src/views/Authorized/Management/bisc/Coworkers/CoworkerCreateView.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { useHistory } from 'react-router-dom'
 import Headline, { SpacingType } from 'components/Chrome/Headline'
 import Actionbar from 'components/Core/Actionbar/Actionbar'
-import { Breadcrumbs } from 'components/Core/Breadcrumb/Breadcrumbs'
+import { Breadcrumbs } from 'components/Core/Breadcrumbs/Breadcrumbs'
 import Button, { ButtonType } from 'components/Core/Button/Button'
 import Input from 'components/Core/DataEntry/Input'
 import { NotificationsManager } from 'components/Core/Feedback/Notifications/NotificationsManager'
@@ -23,7 +23,7 @@ import { EmailValidators } from 'utils/validators/EmailValidators'
 import { GenericValidators } from 'utils/validators/GenericValidators'
 import { FormModel } from './CoworkerOverviewView'
 import { coworkersCreateResponse } from './Detail/coworkers'
-import { breadcrumbItems } from 'components/Core/Breadcrumb/breadcrumbItems'
+import { breadcrumbItems } from 'components/Core/Breadcrumbs/breadcrumbItems'
 
 interface Props {}
