+++ client/src/views/Authorized/Management/taalhuis/Coworkers/Detail/CoworkerReadView.tsx
@@ -9,14 +9,14 @@ import React from 'react'
 import { useHistory } from 'react-router-dom'
 import Headline, { SpacingType } from 'components/Chrome/Headline'
 import Actionbar from 'components/Core/Actionbar/Actionbar'
-import { Breadcrumbs } from 'components/Core/Breadcrumb/Breadcrumbs'
+import { Breadcrumbs } from 'components/Core/Breadcrumbs/Breadcrumbs'
 import Button, { ButtonType } from 'components/Core/Button/Button'
 import { IconType } from 'components/Core/Icon/IconType'
 import Column from 'components/Core/Layout/Column/Column'
 import Space from 'components/Core/Layout/Space/Space'
 import { routes } from 'routes/routes'
 import { ManagementTaalhuisLocationStateProps } from './CoworkersDetailView'
-import { breadcrumbItems } from 'components/Core/Breadcrumb/breadcrumbItems'
+import { breadcrumbItems } from 'components/Core/Breadcrumbs/breadcrumbItems'
 
 interface Props {
     routeState: ManagementTaalhuisLocationStateProps
