+++ client/src/views/Authorized/Participants/taalhuis/Participants/Detail/Documents/ParticipantsDocumentsView.tsx
@@ -2,8 +2,8 @@ import { useLingui } from '@lingui/react'
 import React from 'react'
 import { t } from '@lingui/macro'
 import Headline, { SpacingType } from 'components/Chrome/Headline'
-import { Breadcrumbs } from 'components/Core/Breadcrumb/Breadcrumbs'
-import { breadcrumbItems } from 'components/Core/Breadcrumb/breadcrumbItems'
+import { Breadcrumbs } from 'components/Core/Breadcrumbs/Breadcrumbs'
+import { breadcrumbItems } from 'components/Core/Breadcrumbs/breadcrumbItems'
 
 interface Props {}
